+++ src/main/java/dinkplugin/util/SerializedPet.java
@@ -1,0 +1,9 @@
+package dinkplugin.util;
+
+import lombok.Value;
+
+@Value
+public class SerializedPet {
+    int itemId;
+    String name;
+}