+++ src/main/java/dinkplugin/notifiers/data/LoginNotificationData.java
@@ -1,9 +1,11 @@
 package dinkplugin.notifiers.data;
 
+import dinkplugin.util.SerializedPet;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.List;
 import java.util.Map;
 
 @Value
@@ -21,12 +23,15 @@ public class LoginNotificationData extends NotificationData {
     Progress questCount;
     Progress questPoints;
     SlayerData slayer;
+    @Nullable // requires Chat Commands plugin to be enabled
+    List<SerializedPet> pets;
 
     @Value
     public static class SkillData {
         long totalExperience;
         int totalLevel;
         Map<String, Integer> levels;
+        Map<String, Integer> experience;
     }
 
     @Value