+++ src/main/java/dinkplugin/util/WorldUtils.java
@@ -6,6 +6,7 @@ import lombok.experimental.UtilityClass;
 import net.runelite.api.Actor;
 import net.runelite.api.Client;
 import net.runelite.api.WorldType;
+import net.runelite.api.annotations.Varbit;
 import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.widgets.Widget;
 import net.runelite.api.widgets.WidgetInfo;
@@ -24,6 +25,7 @@ public class WorldUtils {
     private final Set<Integer> CASTLE_WARS_REGIONS = ImmutableSet.of(9520, 9620);
     private final Set<Integer> CLAN_WARS_REGIONS = ImmutableSet.of(12621, 12622, 12623, 13130, 13131, 13133, 13134, 13135, 13386, 13387, 13390, 13641, 13642, 13643, 13644, 13645, 13646, 13647, 13899, 13900, 14155, 14156);
     private final Set<Integer> COX_REGIONS = ImmutableSet.of(12889, 13136, 13137, 13138, 13139, 13140, 13141, 13145, 13393, 13394, 13395, 13396, 13397, 13401);
+    private final Set<Integer> GALVEK_REGIONS = ImmutableSet.of(6486, 6487, 6488, 6489, 6742, 6743, 6744, 6745);
     private final Set<Integer> GAUNTLET_REGIONS = ImmutableSet.of(7512, 7768, 12127); // includes CG
     private final Set<Integer> LMS_REGIONS = ImmutableSet.of(13658, 13659, 13660, 13914, 13915, 13916, 13918, 13919, 13920, 14174, 14175, 14176, 14430, 14431, 14432);
     private final Set<Integer> POH_REGIONS = ImmutableSet.of(7257, 7513, 7514, 7769, 7770, 8025, 8026);
@@ -34,6 +36,16 @@ public class WorldUtils {
     private final int NMZ_REGION = 9033;
     private final int TZHAAR_CAVE = 9551;
     public final @VisibleForTesting int TZHAAR_PIT = 9552;
+
+    /**
+     * @see <a href="https://oldschool.runescape.wiki/w/RuneScape:Varbit/6104">Wiki</a>
+     */
+    private final @Varbit int DRAGON_SLAYER_II_PROGRESS = 6104;
+
+    /**
+     * @see <a href="https://chisel.weirdgloop.org/varbs/display?varbit=6104#ChangeFrequencyTitle">Chisel</a>
+     */
+    private final int DRAGON_SLAYER_II_COMPLETED = 215;
 
     public static WorldPoint getLocation(Client client) {
         return getLocation(client, client.getLocalPlayer());
@@ -77,6 +89,10 @@ public class WorldUtils {
 
     public boolean isClanWars(int regionId) {
         return CLAN_WARS_REGIONS.contains(regionId);
+    }
+
+    public boolean isGalvekRematch(Client client, int regionId) {
+        return GALVEK_REGIONS.contains(regionId) && client.getVarbitValue(DRAGON_SLAYER_II_PROGRESS) >= DRAGON_SLAYER_II_COMPLETED;
     }
 
     public boolean isGauntlet(int regionId) {
@@ -125,7 +141,8 @@ public class WorldUtils {
         }
 
         if (isBarbarianAssault(regionId) || isChambersOfXeric(regionId) || isInferno(regionId) ||
-            isNightmareZone(regionId) || isTzHaarFightCave(regionId) || isPestControl(client)) {
+            isNightmareZone(regionId) || isTzHaarFightCave(regionId) || isPestControl(client) ||
+            isGalvekRematch(client, regionId)) {
             // All PvM activities are dangerous for Hardcore group iron players
             return Utils.getAccountType(client) != AccountType.HARDCORE_GROUP_IRONMAN;
         }