+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Minor: Prefer historical name for first Recipe for Disaster quest. (#352)
+
 ## 1.6.4
 
 - Minor: Add task progress metadata for diary notifications. (#331)