+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Minor: Allow notifications on seasonal worlds to be ignored via advanced config. (#357)
+
 ## 1.6.4
 
 - Minor: Add task progress metadata for diary notifications. (#331)